+++ CHANGELOG.md
@@ -31,6 +31,7 @@
 - Added support for `resample` frequency `W`, `ME`, `YE` with `closed = "left"`.
 - Added support for `DataFrame.rolling.corr` and `Series.rolling.corr` for `pairwise = False` and int `window`.
 - Added support for string time-based `window` and `min_periods = None` for `Rolling`.
+- Added support for `pd.read_sas` (Uses native pandas for processing).
 
 #### Improvements